% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preCompData.R
\docType{data}
\name{allDMCT}
\alias{allDMCT}
\title{Identified DMCTs for each cell type for LIHC DNAm data}
\format{
A list named "allDMCT" with 5 entries. Each entry is a vector of DMCTs for a cell type.
}
\usage{
data(preCompData)
}
\description{
A list, with each entry for DMCTs of each cell type
}
\keyword{datasets}
