% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preCompData.R
\docType{data}
\name{clustHeatmap}
\alias{clustHeatmap}
\title{A heatmap object of standardized residual matrix over lymphocyte DMCTs}
\format{
A \code{\link[ComplexHeatmap]{Heatmap-class}} object
}
\usage{
data(preCompData)
}
\description{
A \code{\link[ComplexHeatmap]{Heatmap-class}} object, for a heatmap of
standardized residual matrix over lymphocyte specific DMCTs, annotated by CELTYC clustering results
using lymphocyte specific DMCTs, Hep specific DMCTs and EC specific DMCTs respectively.
}
\keyword{datasets}
