% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenResidualMat.R
\name{GenResidualMat}
\alias{GenResidualMat}
\title{Generate a residual matrix adjusted for cell type fractions from the original bulk DNAm data}
\usage{
GenResidualMat(beta.m, estCTF.m, standardize = T, ncores = 4)
}
\arguments{
\item{beta.m}{A beta value matrix, with rows representing CpG sites and columns representing samples.}

\item{estCTF.m}{A matrix of cell-type fractions, with columns representing cell types and rows representing samples.}

\item{standardize}{A boolean value to determine whether to apply z-score standardization (zero-mean and unit-variance) to the residual matrix.
Default is TRUE.}

\item{ncores}{The number of cores to use, i.e. at most how many child processes will be run simultaneously.}
}
\value{
A (standardized if the parameter "standardized" is set to TRUE) residual matrix.
}
\description{
This is a function to regress out estimated cell type fractions from the original DNAm data,
i.e. the beta value matrix.
}
\examples{
\dontrun{
data("LIHC_data")
print(sd(LIHC_data$DNAm[1,]))
res.m <- GenResidualMat(LIHC_data$DNAm,estCTF.m = LIHC_data$estF,standardize = T,ncores = 40)
print(sd(res.m[1,]))
}

}
