% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetDMCT.R
\name{GetDMCT}
\alias{GetDMCT}
\title{Extract DMCT sets from the output of CellDMC method}
\usage{
GetDMCT(
  celldmc.o,
  nDMCT = 100,
  specific = T,
  nSpDMCT = 100,
  common = T,
  nCT = 3,
  nCmDMCT = 100
)
}
\arguments{
\item{celldmc.o}{A list object, which is the output of the function \code{\link[EpiDISH]{CellDMC}}}

\item{nDMCT}{An integer value, which acts as a threshold value.
For a cell type, if the number of its DMCTs is smaller than nDMCT, the vector of its DMCTs will not be saved in the "dmct" entry of the returned list.}

\item{specific}{Logical value, indicating whether to select DMCTs specific to each cell type or not. Default is TRUE.}

\item{nSpDMCT}{An integer value, which acts as a threshold value.
For a cell type, if the number of the DMCTs that are specific to that cell type is smaller than nSpDMCT, the vector of its specific DMCTs will not be saved in the "spec" entry of the returned list.}

\item{common}{Logical value, indicating whether to select DMCTs which are common to different cell types. Default is TRUE.}

\item{nCT}{Integer value. The number of cell types for users to select common DMCTs.}

\item{nCmDMCT}{An integer value, which acts as a threshold value.
For a combination of nCT cell types, if the number of the DMCTs common to these cell types is smaller than nCmDMCT, the vector of these common DMCTs will not be saved in the "comm" entry of the returned list.}
}
\value{
A list with the following entries:
\itemize{
\item dmct: A list containing the vectors of DMCTs altered in different cell types.
Only DMCT vectors with a length >= nDMCT will be saved in this entry.
\item spec: A list containing the vectors of DMCTs exclusively altered in each of different cell types.
Only DMCT vectors with a length >= nSpDMCT will be saved in this entry. Note that this entry will be NULL if parameter "specific" is set to be FALSE.
\item comm: A list containing the vectors of DMCTs common to nCT cell types.
Only DMCT vectors with a length >= nCmDMCT will be saved in this entry. Note that this entry will be NULL if parameter "common" is set to be FALSE.
}
}
\description{
This function helps extract disease/phenotype associated differentially methylated CpG sites altered in one or more cell types (i.e. DMCTs)
from the output of the function  EpiDISH::CellDMC (\code{\link[EpiDISH]{CellDMC}}).

It allows the users to retrieve DMCTs which are:
\enumerate{
\item Altered in each cell type (not necessarily the DMCTs specific to only one cell type).
\item Exclusively altered in each cell type.
\item Common to multiple cell types.
}
}
\details{
The retrieved DMCT sets can serve as the input of \code{\link[CELTYC]{DoCELTYC}} function.
Note that in order to do cell-type specific clustering properly, one should use a reasonable number of DMCTs.
Therefore, if there are not enough DMCTs for a particular cell type (fewer than the pre-set threshold parameter nDMCT),
getDMCT will not save that vector of DMCTs altered in that cell type.
Similarly, when users try to get DMCTs specific to a particular cell type (with the parameter "specific" set to TRUE),
if the number of the specific DMCTs is smaller than nSpDMCT,
getDMCT will not save that vector of DMCTs exclusively altered in that cell type.
When users try to get DMCTs common to nCT cell types (with the parameter "common" set to TRUE),
given a combination of nCT cell types,
if the number of the common DMCTs is smaller than nCmDMCT,
getDMCT will not save that vector of DMCTs common to these cell types.

In general, these parameters (nDMCT, nSpDMCT, nCmDMCT) help users to keep only the DMCT sets containing a satisfying number of DMCTs.
}
