% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetResidualMat.R
\name{GetResidualMat}
\alias{GetResidualMat}
\title{Generate a residual matrix adjusted for cell type fractions from the original bulk DNAm data}
\usage{
GetResidualMat(beta.m, estCTF.m, standardize = T, ncores = 4)
}
\arguments{
\item{beta.m}{A beta matrix, with rows representing CpG sites and columns representing samples.}

\item{estCTF.m}{A matrix of cell type fractions, with columns representing cell types and rows representing samples.}

\item{standardize}{A boolean value to determine whether to do standardization (zero-mean and unit-variance) for the residual matrix.
Default is TRUE.}

\item{ncores}{The number of cores to use, i.e. at most how many child processes will be run simultaneously.}
}
\value{
A (standardized) residual matrix.
}
\description{
This is a function to regress out estimated cell type fractions from the original DNAm data,
i.e. the beta matrix.
}
