% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preCompData.R
\docType{data}
\name{LIHC_DNAm}
\alias{LIHC_DNAm}
\title{LIHC DNAm beta matrix}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 4744 rows and 379 columns.
}
\usage{
data(preCompData)
}
\description{
A matrix, with rows for cancer-associated DMCs altered in Lymphocytes (identified with \code{\link[EpiDISH]{CellDMC}}),
columns for samples of cancer status
}
\keyword{datasets}
