% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preCompData.R
\docType{data}
\name{LIHC_estF}
\alias{LIHC_estF}
\title{Estimated CTFs for LIHC DNAm data (for cancer status samples)}
\format{
A matrix
}
\usage{
data(preCompData)
}
\description{
A matrix, with rows for samples and columns for cell types
}
\keyword{datasets}
