% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preCompData.R
\docType{data}
\name{LIHC_pheno}
\alias{LIHC_pheno}
\title{Phenotype information for LIHC DNAm data (for cancer status samples)}
\format{
A matrix
}
\usage{
data(preCompData)
}
\description{
A matrix, with rows for samples and columns for different phenotypes
}
\keyword{datasets}
