% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preCompData.R
\docType{data}
\name{selDMCT}
\alias{selDMCT}
\title{Subsets of DMCTs for each cell type for LIHC DNAm data}
\format{
An object of class \code{list} of length 4.
}
\usage{
data(preCompData)
}
\description{
A list, with each entry for DMCTs of each cell type.
For entries named "Lym", "Hep", "EC": these DMCTs are non-overlapped for these cell types.
For entry "common", these are DMCTs shared by 3 cell types.
}
\keyword{datasets}
